+++ CHANGELOG.md
@@ -64,6 +64,7 @@
 - Added `amend` function which amends a new object into an old to update data
 - Added `clamp` function which clamps a number between 2 others
 - Added `defaultTo` function which returns a value unless it is `NaN`, `undefined`, or `null` then it returns a default
+- Added `reduceRight` function which works like `reduce` but starts from the right of the array
 
 ## v0.9.2